Use CoronaTestRestoreHandler and CoronaTestQRCodeHandler
