+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/qrcode/ui/QrCodeScannerViewModel.kt
@@ -5,7 +5,6 @@ import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import de.rki.coronawarnapp.coronatest.qrcode.CoronaTestQRCode
 import de.rki.coronawarnapp.coronatest.type.CoronaTest
-import de.rki.coronawarnapp.coronatest.type.PersonalCoronaTest
 import de.rki.coronawarnapp.covidcertificate.common.certificate.DccMaxPersonChecker
 import de.rki.coronawarnapp.covidcertificate.common.qrcode.DccQrCode
 import de.rki.coronawarnapp.covidcertificate.common.repository.CertificateContainerId
@@ -17,22 +16,20 @@ import de.rki.coronawarnapp.presencetracing.TraceLocationSettings
 import de.rki.coronawarnapp.presencetracing.checkins.qrcode.CheckInQrCode
 import de.rki.coronawarnapp.qrcode.QrCodeFileParser
 import de.rki.coronawarnapp.qrcode.handler.CheckInQrCodeHandler
+import de.rki.coronawarnapp.qrcode.handler.CoronaTestQRCodeHandler
 import de.rki.coronawarnapp.qrcode.handler.DccQrCodeHandler
 import de.rki.coronawarnapp.qrcode.parser.QrCodeBoofCVParser
 import de.rki.coronawarnapp.qrcode.scanner.ImportDocumentException
 import de.rki.coronawarnapp.qrcode.scanner.ImportDocumentException.ErrorCode.CANT_READ_FILE
 import de.rki.coronawarnapp.qrcode.scanner.QrCodeValidator
-import de.rki.coronawarnapp.reyclebin.coronatest.RecycledCoronaTestsProvider
-import de.rki.coronawarnapp.reyclebin.coronatest.request.toRestoreRecycledTestRequest
+import de.rki.coronawarnapp.reyclebin.coronatest.handler.CoronaTestRestoreEvent
+import de.rki.coronawarnapp.reyclebin.coronatest.handler.CoronaTestRestoreHandler
 import de.rki.coronawarnapp.reyclebin.covidcertificate.RecycledCertificatesProvider
-import de.rki.coronawarnapp.submission.SubmissionRepository
 import de.rki.coronawarnapp.tag
-import de.rki.coronawarnapp.util.HashExtensions.toSHA256
 import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
 import de.rki.coronawarnapp.util.ui.SingleLiveEvent
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModel
 import de.rki.coronawarnapp.util.viewmodel.SimpleCWAViewModelFactory
-import kotlinx.coroutines.flow.first
 import timber.log.Timber
 
 @Suppress("LongParameterList")
@@ -43,11 +40,11 @@ class QrCodeScannerViewModel @AssistedInject constructor(
     private val dccHandler: DccQrCodeHandler,
     private val checkInHandler: CheckInQrCodeHandler,
     private val dccTicketingQrCodeHandler: DccTicketingQrCodeHandler,
-    private val submissionRepository: SubmissionRepository,
+    private val coronaTestQRCodeHandler: CoronaTestQRCodeHandler,
+    private val coronaTestRestoreHandler: CoronaTestRestoreHandler,
     private val dccSettings: CovidCertificateSettings,
     private val traceLocationSettings: TraceLocationSettings,
     private val recycledCertificatesProvider: RecycledCertificatesProvider,
-    private val recycledCoronaTestsProvider: RecycledCoronaTestsProvider,
     private val dccMaxPersonChecker: DccMaxPersonChecker
 ) : CWAViewModel(dispatcherProvider) {
 
@@ -125,36 +122,16 @@ class QrCodeScannerViewModel @AssistedInject constructor(
         result.postValue(containerId.toDccDetails())
     }
 
-    fun restoreCoronaTest(recycledCoronaTest: CoronaTest) =
-        when (recycledCoronaTest) {
-            is PersonalCoronaTest -> restorePersonalCoronaTest(recycledCoronaTest)
-            else -> { /* to do */ }
-        }
-
-    private fun restorePersonalCoronaTest(recycledCoronaTest: PersonalCoronaTest) = launch {
-        val currentCoronaTest = submissionRepository.testForType(recycledCoronaTest.type).first()
-        when {
-            currentCoronaTest != null -> CoronaTestResult.RestoreDuplicateTest(
-                recycledCoronaTest.toRestoreRecycledTestRequest()
-            )
-
-            else -> {
-                recycledCoronaTestsProvider.restoreCoronaTest(recycledCoronaTest.identifier)
-                recycledCoronaTest.toCoronaTestResult()
-            }
-        }.also {
-            result.postValue(it)
-        }
+    fun restoreCoronaTest(recycledCoronaTest: CoronaTest) = launch {
+        val coronaTestRestoreEvent = coronaTestRestoreHandler.restoreCoronaTest(recycledCoronaTest)
+        result.postValue(coronaTestRestoreEvent.toCoronaResult())
     }
 
-    private fun PersonalCoronaTest.toCoronaTestResult(): CoronaTestResult = when {
-        isPending -> CoronaTestResult.TestPending(test = this)
-        isNegative -> CoronaTestResult.TestNegative(test = this)
-        isPositive -> when (isAdvancedConsentGiven) {
-            true -> CoronaTestResult.TestPositive(test = this)
-            false -> CoronaTestResult.WarnOthers(test = this)
-        }
-        else -> CoronaTestResult.TestInvalid(test = this)
+    private fun CoronaTestRestoreEvent.toCoronaResult(): CoronaTestResult = when (this) {
+        is CoronaTestRestoreEvent.RestoreDuplicateTest -> CoronaTestResult.RestoreDuplicateTest(
+            restoreRecycledTestRequest
+        )
+        else -> TODO()
     }
 
     private suspend fun onDccQrCode(dccQrCode: DccQrCode) {
@@ -197,14 +174,7 @@ class QrCodeScannerViewModel @AssistedInject constructor(
     }
 
     private suspend fun onCoronaTestQrCode(qrCode: CoronaTestQRCode) {
-        Timber.tag(TAG).d("onCoronaTestQrCode()")
-        val recycledCoronaTest = recycledCoronaTestsProvider.findCoronaTest(qrCode.rawQrCode.toSHA256())
-
-        val coronaTestResult = when {
-            recycledCoronaTest != null -> CoronaTestResult.InRecycleBin(recycledCoronaTest)
-            else -> CoronaTestResult.TestRegistrationSelection(qrCode)
-        }
-        Timber.tag(TAG).d("coronaTestResult=${coronaTestResult::class.simpleName}")
+        val coronaTestResult = coronaTestQRCodeHandler.handleQrCode(qrCode)
         result.postValue(coronaTestResult)
     }
